+++ src/services/school/index.js
@@ -1,5 +1,5 @@
 const service = require('feathers-mongoose');
-const { Configuration } = require('@schul-cloud/commons');
+const { Configuration } = require('@hpi-schul-cloud/commons');
 const { static: staticContent } = require('@feathersjs/express');
 const path = require('path');