+++ src/utils/apiValidation.js
@@ -1,5 +1,5 @@
 const OpenApiValidator = require('express-openapi-validator');
-const commons = require('@schul-cloud/commons');
+const commons = require('@hpi-schul-cloud/commons');
 
 const { Configuration } = commons;